Get coverage back to 100%
